Refactoring: let `review` to accept only `argv`.
Until now, `code-review-gpt` assumes that the review is local accessible
code, and so it can figure out the files to review via git.

But this assumption is going to be changed with this enhancement: remote
code review.

If passing `--remote` CLI option, `code-review-gpt` will use a different
strategy to get the files to be reviewed.